- cherrypicked modifications regarding LoRaWAN and module operation stuff (not the low-level stuff) from 1.0.6RC8 and applied to 1.0.5

+++ software/firmware/RN2903NA_1.0.5/Source/RN2903_Parser/parser/parser_radio.c
@@ -584,7 +584,7 @@ void Parser_RadioGetSnr(parserCmdInfo_t* pParserCmdInfo)
 
 void Parser_RadioGetRSSI(parserCmdInfo_t* pParserCmdInfo)
 {
-    int8_t packetRSSI;
+    int16_t packetRSSI;
 
     packetRSSI = RADIO_GetPacketRSSI();
     itoa(aParserData, packetRSSI, 10U);
